backend/pyproject.toml
[project]
name = "kitchencrew"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "arize-phoenix-otel>=0.9.2",
    "beautifulsoup4>=4.12.0",
    "click>=8.2.1",
    "crewai>=0.41.0",
    "crewai-tools>=0.8.0",
    "fastapi>=0.104.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.1.0",
    "openai>=1.75.0",
    "openinference-instrumentation-crewai>=0.1.9",
    "openinference-instrumentation-langchain>=0.1.43",
    "openinference-instrumentation-litellm>=0.1.19",
    "pandas>=2.1.0",
    "pydantic>=2.0.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "rich>=13.9.4",
    "typing-extensions>=4.8.0",
    "uvicorn>=0.24.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
]
